+++ README.md
@@ -11,7 +11,7 @@ Clone the Git repository and run the following commands:
 ```sh
 npm install
 export DISCORD_TOKEN="<your Discord bot token>"
-export GOOGLE_TOKEN="<your Google Maps API key>"
+export GOOGLE_TOKEN="<your Google (Maps) API key>"
 export CR_ADDRESS="http://media.wizards.com/2016/docs/MagicCompRules_20160930.txt"
 export IPG_ADDRESS="https://sites.google.com/site/mtgfamiliar/rules/InfractionProcedureGuide-light.html"
 export MTR_ADDRESS="https://sites.google.com/site/mtgfamiliar/rules/MagicTournamentRules-light.html"
